private/storage/streams/buffer: add chunk backend
ChunkBackend is an alternative to MemoryBackend. Instead of one
large allocation that can hold the maximum encrypted size for each
segment, it uses 1MiB chunks. These chunks are similarly allocated from a
sync.Pool to reduce allocation pressure.

The motivation is to drastically reduce the base memory requirement for
an individual upload containing partial segments. This should allow edge
services to support an increase in concurrent uploads.

The splitter is updated to use this backend in lieu of the memory
backend.

Change-Id: I6b33beffc22b3a7d5525a010a36bb2ae58de0cb3