private/stream.Download: retry from current offset after decryption error
This change is a short term fix for piece corruption. It uses the
observation that piece corruption is extremely rare and each time the
Satellite returns a set of pieces, it doesn't return all of them.
If we get an authenticated decryption failure, it's likely that just
asking the satellite for a different collection of pieces a few times
will likely skip the bad piece once or twice.

In the long term we want to change all of the plumbing of the uplink
to pass authenticated decryption machinery back into the RS code so
the RS code knows to try and correct from errors, but we maybe don't
need to do that at current error rates.

Either way, we will still never return corrupted data to clients
thanks to authenticated encryption, so trying again at the same
offset (this change) is always safe.

Change-Id: I6a6db735a6521ee02d172e89baec90cc38a61ded