private/storage/streams: use satellite's limit choice directly
prior to this change, the uplink attempts to enforce its own
idea of what the download tail tolerance should be. however,
the satellite should be in charge of this (and in fact
already is). so this change makes that official and
just uses the limits the satellite provides directly (which
is already the right amount).

see https://review.dev.storj.io/c/storj/common/+/12257

Change-Id: I86754897342b071938919ac8d3a0386fb875596e