private/metaclient: implement CompressedBatch and use it
this causes some rpcs to use Batch by defualt instead
of their direct call variants. this will allow them
to be compressed if the STORJ_COMPRESSED_BATCH env
variable is set (soon to be default in the future)
and the satellite supports the rpc.

there should be no change in behavior until then.

Change-Id: I7b027ffea217fa90e463cef58ff1d4ed52189375