move/copy: support objects that end with a trailing slash
unlike a filesystem, where a slash is a special character,
there are no restrictions or requirements about object keys
in object storage. an object can end with a trailing slash,
and listing can choose a delimiter other than a slash.

so, as a result, it is insufficient to test if a user
is trying to do a recursive copy or move by checking if
the input ends with a slash. they may genuinely have
an object with a trailing slash in the object name, or
may want one, and we need to support move and copy with
those objects.

see https://github.com/storj/customer-issues/issues/2066#issuecomment-2261252900
for more

Change-Id: I2cdc581e19d6bc1a9b47b3b4f7bae1ceb3cd1446