private/metaclient: Fix multipart upload list
ListPendingObjecStreams didn't pass set the Cursor to the returned list
object lists, so when the first page didn't have all the items the
request for a next page didn't carry the cursor and the server returned
the first page.

This commit fixes the issue and add tests cases to verify the specific
case of beginning an object upload with exactly the same key and how the
list upload method return the same number of begun uploads.

Change-Id: I84128adb6b8f4b1f51adc63d3213d36afda72822