private/metaclient: support lite mode for BeginSegment
most of the ugliness is limited to the cursed
package which is mostly vendored from the
satellite so that the uplink can fill in satellite
signatures and repack the segment id if it has a
satellite identity. this lets the satellite
respond with order limits excluding the high
entropy piece id and satellite signature fields,
letting compression kick in much better. i saw a
47% reduction in BeginSegment compressed response
sizes in storj-sim.

it currently doesn't have any way to set the
satellite field on the client, so the code
will be a no-op.

Change-Id: I18a6f68770f53a785e3bfb738992b569d75bdbbd