private/object: prevent unencryptable key from failing DeleteObjects
This change ensures that DeleteObjects isn't interrupted when an object
key that cannot be encrypted is provided. Previously, when such a key
was encountered, we would immediately return an error, and the
DeleteObjects request would never reach the satellite. Now, we record a
failure for that key and continue the operation.

This issue occured when a provided key violated a prefix restriction
placed on the access grant that was used when opening the project.

Change-Id: If427f8ee2d346fdd141827c08d181674f109ad59