private/stream: enable error detection for specific offset
Current code is enabling error detection only for first object
segment. This makes download retries inefficient because then
downloaded data is not going through correction and in case of
malformed piece there is high chance to fail download because of
decryption error.

With this change on download retry error detection will be enabled
for specific segment that was problematic before retry.

Change-Id: Id05f0427df63cee409b51f3fb63367c0713f4bad