private/stream: retry segment download on network failure
This is initial attempt to avoid failing downloads in case network
errors. This change is reusing logic we have for decryption and
quiescence errors and in case of network error it retries downloading
whole segment.

List of errors that will be retried is rather short but will be
updated if we will find cases that should be covered too.

Change-Id: I385d0163465181ef7e013e4e5cf1f59e1654cc2f